Work on  Cellular Automata and spawning players in
When Play is pressed it starts up Level1.cs which inside contains a tilemap class which creates a cellular automata configuration. I've added some code that randomly spawns the player anywhere there is a floor. Algorithm is currently buggy an need to find a way of eliminating rooms that seperate from the cave. Collisions are yet to be added to this tilemap, will do that tomorrow night after work.